event_flyer_generator/requirements.txt
flask
openai>=1.0.0
jinja2
python-dotenv
